Extracted e2e monitoring tests (#3973)
* Extracted e2e monitoring tests
