+++ .github/workflows/monitoring-claim-orders .yml
@@ -25,13 +25,13 @@ jobs:
             ${{ runner.OS }}-20.x-
       - name: Install Playwright
         run: |
-          yarn --cwd packages/web install --frozen-lockfile && npx playwright install --with-deps chromium
+          yarn --cwd packages/e2e install --frozen-lockfile && npx playwright install --with-deps chromium
       - name: Run Swap tests in US
         env:
           BASE_URL: "https://app.osmosis.zone"
           PRIVATE_KEY: ${{ secrets.TEST_PRIVATE_KEY_1 }}
         run: |
-          cd packages/web
+          cd packages/e2e
           npx playwright test claim
       - name: upload test results
         if: failure()
@@ -39,7 +39,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: claim1-test-results
-          path: packages/web/playwright-report
+          path: packages/e2e/playwright-report
 
   fe-claim2-tests:
     timeout-minutes: 10
@@ -60,13 +60,13 @@ jobs:
             ${{ runner.OS }}-20.x-
       - name: Install Playwright
         run: |
-          yarn --cwd packages/web install --frozen-lockfile && npx playwright install --with-deps chromium
+          yarn --cwd packages/e2e install --frozen-lockfile && npx playwright install --with-deps chromium
       - name: Run Claim tests
         env:
           BASE_URL: "https://app.osmosis.zone"
           PRIVATE_KEY: ${{ secrets.TEST_PRIVATE_KEY_2 }}
         run: |
-          cd packages/web
+          cd packages/e2e
           npx playwright test claim
       - name: upload claim2 test results
         if: failure()
@@ -74,7 +74,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: claim2-test-results
-          path: packages/web/playwright-report
+          path: packages/e2e/playwright-report
 
   delete-deployments:
     runs-on: ubuntu-latest
