+++ .github/workflows/monitoring-e2e-tests.yml
@@ -114,12 +114,12 @@ jobs:
             ${{ runner.OS }}-20.x-
       - name: Install Playwright
         run: |
-          yarn --cwd packages/web install --frozen-lockfile && npx playwright install --with-deps chromium
+          yarn --cwd packages/e2e install --frozen-lockfile && npx playwright install --with-deps chromium
       - name: Run Swap tests on ${{ matrix.env }}
         env:
           BASE_URL: ${{ matrix.base-url }}
         run: |
-          cd packages/web
+          cd packages/e2e
           npx playwright test pools
       - name: upload test results
         if: failure()
@@ -127,7 +127,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: ${{ matrix.env }}-quote-test-results
-          path: packages/web/playwright-report
+          path: packages/e2e/playwright-report
 
   delete-deployments:
     runs-on: ubuntu-latest
