+++ .github/workflows/prod-frontend-e2e-tests.yml
@@ -52,14 +52,14 @@ jobs:
             ${{ runner.OS }}-20.x-
       - name: Install Playwright
         run: |
-          yarn --cwd packages/web install --frozen-lockfile && npx playwright install --with-deps chromium
+          yarn --cwd packages/e2e install --frozen-lockfile && npx playwright install --with-deps chromium
       - name: Run Swap Pair tests on Master
         env:
           BASE_URL: "https://app.osmosis.zone"
           PRIVATE_KEY: ${{ secrets.TEST_PRIVATE_KEY }}
           WALLET_ID: ${{ secrets.TEST_WALLET_ID }}
         run: |
-          cd packages/web
+          cd packages/e2e
           npx playwright test transactions portfolio swap.wallet
       - name: upload test results
         if: always()
@@ -67,7 +67,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: prod-e2e-test-results-${{ github.run_id }}
-          path: packages/web/playwright-report
+          path: packages/e2e/playwright-report
       - name: Send Slack alert if test fails
         id: slack
         if: failure()
@@ -124,12 +124,12 @@ jobs:
             ${{ runner.OS }}-20.x-
       - name: Install Playwright
         run: |
-          yarn --cwd packages/web install --frozen-lockfile && npx playwright install --with-deps chromium
+          yarn --cwd packages/e2e install --frozen-lockfile && npx playwright install --with-deps chromium
       - name: Run Pools tests
         env:
           BASE_URL: "https://app.osmosis.zone"
         run: |
-          cd packages/web
+          cd packages/e2e
           npx playwright test pools select
       - name: upload pools test results
         if: failure()
@@ -137,4 +137,4 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: prod-pools-test-results-${{ github.run_id }}
-          path: packages/web/playwright-report
+          path: packages/e2e/playwright-report
