+++ package.json
@@ -29,7 +29,7 @@
     "pre-commit": "lerna run pre-commit --concurrency false",
     "publish:libs": "lerna publish",
     "machine-translate": "inlang machine translate -f --project ./project.inlang && yarn lint:fix",
-    "lint:workspace": "npx sherif@latest --ignore-package ./packages/keplr-hooks --ignore-package ./packages/keplr-stores",
+    "lint:workspace": "npx sherif@latest --ignore-package ./packages/keplr-hooks --ignore-package ./packages/keplr-stores --ignore-package ./packages/e2e",
     "lint:workspace:fix": "yarn lint:workspace -f",
     "postinstall": "yarn lint:workspace"
   },
