+++ packages/e2e/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "app-osmosis-e2e",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "clean": "rm -rf node_modules; rm -rf .next",
+    "lint": "biome check --write",
+    "lint:fix": "biome check --fix --unsafe",
+    "test": "npx playwright test",
+    "setup": "yarn install && npx playwright install --with-deps chromium"
+  },
+  "dependencies": {
+    "@playwright/test": "1.49.0",
+    "decompress": "4.2.1",
+    "dotenv": "16.4.5"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.9.4",
+    "@types/node": "22.7.5",
+    "typescript": "5.6.3"
+  }
+}
