+++ packages/e2e/playwright.config.ts
@@ -1,0 +1,41 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import { defineConfig, devices } from "@playwright/test";
+
+import "dotenv/config";
+
+/**
+ * See https://playwright.dev/docs/test-configuration.
+ */
+export default defineConfig({
+  /* Reporter to use. See https://playwright.dev/docs/test-reporters */
+  reporter: [
+    ["html", { open: "never" }],
+    ["junit", { outputFile: "./playwright-report/test-results.xml" }],
+  ],
+  timeout: 45000,
+  testDir: "./tests",
+  /* Run tests in files in parallel. */
+  fullyParallel: true,
+  /* Fail the build on CI if you accidentally left test.only in the source code. */
+  forbidOnly: !!process.env.CI,
+  /* Retry on CI only */
+  retries: process.env.CI ? 1 : 0,
+  /* Opt out of parallel tests on CI. Multiple swaps on same account could be unsatble. Test execution will be batched per agent. */
+  workers: 1,
+  /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
+  use: {
+    baseURL: process.env.BASE_URL ?? "https://stage.osmosis.zone",
+
+    /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
+    trace: "off",
+    screenshot: "only-on-failure",
+  },
+
+  /* Configure projects for major browsers */
+  projects: [
+    {
+      name: "chromium",
+      use: { ...devices["Desktop Chrome"] },
+    },
+  ],
+});
