+++ packages/e2e/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "compilerOptions": {
+    "target": "ES6", // Specify ECMAScript target version
+    "module": "commonjs", // Specify module code generation
+    "strict": true, // Enable all strict type-checking options
+    "esModuleInterop": true, // Enables emit interoperability between CommonJS and ES Modules
+    "skipLibCheck": true, // Skip type checking of declaration files
+    "forceConsistentCasingInFileNames": true, // Disallow inconsistently-cased references to the same file
+    "types": ["node"] // Include Node.js type definitions
+  },
+  "include": [
+    "pages/**/*",
+    "tests/**/*",
+    "test-config.ts",
+    "unzip-extension.ts",
+    "playwright.config.ts"
+  ],
+  "exclude": ["node_modules"]
+}
