+++ packages/web/package.json
@@ -95,7 +95,6 @@
     "classnames": "^2.3.2",
     "dayjs": "^1.10.7",
     "debounce": "^1.2.1",
-    "decompress": "^4.2.1",
     "eventemitter3": "^4.0.7",
     "fancy-canvas": "^2.1.0",
     "fast-deep-equal": "^3.1.3",
@@ -135,7 +134,6 @@
   },
   "devDependencies": {
     "@keplr-wallet/provider-mock": "0.12.12",
-    "@playwright/test": "^1.48.0",
     "@testing-library/jest-dom": "^6.1.4",
     "@testing-library/react": "^14.1.2",
     "@testing-library/user-event": "^14.5.1",
@@ -162,7 +160,6 @@
     "jest-in-case": "^1.0.2",
     "jest-launchdarkly-mock": "^2.1.0",
     "knip": "^5.17.4",
-    "monocart-reporter": "^2.6.0",
     "msw-trpc": "1.3.3",
     "postcss": "^8.4.5",
     "prettier": "^2.8.8",
