--- packages/web/playwright.config.ts
@@ -1,52 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { defineConfig, devices } from "@playwright/test";
-
-/**
- * Read environment variables from file.
- * https://github.com/motdotla/dotenv
- */
-// require('dotenv').config();
-
-/**
- * See https://playwright.dev/docs/test-configuration.
- */
-export default defineConfig({
-  /* Reporter to use. See https://playwright.dev/docs/test-reporters */
-  reporter: [
-    ["html", { open: "never" }],
-    ["junit", { outputFile: "./playwright-report/test-results.xml" }],
-    [
-      "monocart-reporter",
-      {
-        name: "Test Coverage Report",
-        outputFile: "./playwright-report/cov-report.html",
-      },
-    ],
-  ],
-  timeout: 45000,
-  testDir: "./e2e/tests",
-  /* Run tests in files in parallel */
-  fullyParallel: false,
-  /* Fail the build on CI if you accidentally left test.only in the source code. */
-  forbidOnly: !!process.env.CI,
-  /* Retry on CI only */
-  retries: process.env.CI ? 2 : 0,
-  /* Opt out of parallel tests on CI. Multiple swaps on same account could be unsatble. Test execution will be batched per agent. */
-  workers: process.env.CI ? 1 : undefined,
-  /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
-  use: {
-    baseURL: process.env.BASE_URL ?? "https://stage.osmosis.zone",
-
-    /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-    trace: "off",
-    screenshot: "only-on-failure",
-  },
-
-  /* Configure projects for major browsers */
-  projects: [
-    {
-      name: "chromium",
-      use: { ...devices["Desktop Chrome"] },
-    },
-  ],
-});
